requirements.txt
# pygame
pygame==2.6.0
pyopengl==3.1.9
trimesh==4.6.8
pyglet==2.1.5
flask==3.1.0